+++ software/main/communication/bluetooth_serial.cpp
@@ -41,8 +41,9 @@ static const esp_spp_mode_t esp_spp_mode = ESP_SPP_MODE_CB;
 
 static const esp_spp_sec_t sec_mask = ESP_SPP_SEC_AUTHENTICATE;
 static const esp_spp_role_t role_slave = ESP_SPP_ROLE_SLAVE;
+static struct timeval time_old;
 
-// static struct timeval time_new, time_old;
+// static struct timeval time_new;
 // static long data_num = 0;
 
 // static void print_speed(void)
