+++ software/main/main.cmake
@@ -1,0 +1,14 @@
+# *************************************************************************
+#
+# Copyright (c) 2022 Andrei Gramakov. All rights reserved.
+#
+# This file is licensed under the terms of the MIT license.  
+# For a copy, see: https://opensource.org/licenses/MIT
+#
+# site:    https://agramakov.me
+# e-mail:  mail@agramakov.me
+#
+# *************************************************************************
+
+file (GLOB SRC_MAIN "${CMAKE_CURRENT_LIST_DIR}/*.c" 
+                    "${CMAKE_CURRENT_LIST_DIR}/*.cpp")
