+++ software/main/motion/motors_on_esp32.cpp
@@ -19,7 +19,7 @@
 #include "soc/mcpwm_periph.h"
 
 #include "config.h"
-#include "delay.h"
+#include "port/delay.h"
 #include "macros.h"
 #include "motors_on_esp32.hpp"
